This commit is a work in progress. There is a segfault that was left in intentionally to implement more debugging features.

+++ src/utils/hash_table.c
@@ -256,7 +256,7 @@ const char* iterate_hash_table(hash_table_t* tab, int reset) {
     if(reset)
         ht_index = -1;
 
-    for(ht_index++; ht_index < tab->capacity; ht_index++) {
+    for(ht_index++; ht_index < (int)tab->capacity; ht_index++) {
         if(tab->entries[ht_index].key != NULL) {
             return tab->entries[ht_index].key;
         }
